saver

+++ LagouSpider/src/main/java/wuxian/me/lagouspider/model/Area.java
@@ -6,7 +6,7 @@ import static com.google.common.base.Preconditions.*;
  * Created by wuxian on 1/4/2017.
  */
 public class Area {
-    public long area_id = -1;
+    public int area_id = -1;
     public String name;
     public String distinct_name;
